Move `mixin.js` to `/src`
