+++ shape_circle.js
@@ -1,6 +1,6 @@
 var Fabric = require('fabric').fabric;
 var extend = Fabric.util.object.extend;
-var mixin = require('./mixin');
+var mixin = require('./src/mixin');
 
 var Circle = Fabric.util.createClass(Fabric.Circle, {
    // Inherited variables with new values.
