+++ shape_line.js
@@ -1,6 +1,6 @@
 var Fabric = require('fabric').fabric;
 var isNode = typeof window == 'undefined';
-var mixin = require('./mixin');
+var mixin = require('./src/mixin');
 
 var Line = Fabric.util.createClass(Fabric.Line, {
    type: 'line',
